requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
gtts
transformers
torch
openai
